Move to lsst namespace
