+++ .gitignore
@@ -3,7 +3,7 @@ build
 builddir
 build-debug
 build-release
-python/gauss2d.egg-info/
+python/lsst/lsst_gauss2d.egg-info/
 
 # IDE folders and files
 .cache/
@@ -20,6 +20,7 @@ compile_commands.json
 *.lo
 *.o
 *.obj
+*.pyc
 
 # Precompiled Headers
 *.gch
