+++ include/meson.build
@@ -1,12 +1,15 @@
+prefix_lsst = 'lsst/'
+prefix_gauss2d = prefix_lsst + 'gauss2d/'
 gauss2d_headers = [
-    'centroid.h',
-    'coordinatesystem.h',
-    'ellipse.h',
-    'evaluate.h',
-    'gaussian.h',
-    'image.h',
-    'object.h',
-    'type_name.h',
-    'vectorimage.h',
+    prefix_lsst + 'gauss2d.h',
+    prefix_gauss2d + 'centroid.h',
+    prefix_gauss2d + 'coordinatesystem.h',
+    prefix_gauss2d + 'ellipse.h',
+    prefix_gauss2d + 'evaluate.h',
+    prefix_gauss2d + 'gaussian.h',
+    prefix_gauss2d + 'image.h',
+    prefix_gauss2d + 'object.h',
+    prefix_gauss2d + 'type_name.h',
+    prefix_gauss2d + 'vectorimage.h',
 ]
-install_headers(gauss2d_headers, subdir : 'gauss2d')
+install_headers(gauss2d_headers, preserve_path : true)
