+++ python/lib/tests/test_pyimage.cc
@@ -4,28 +4,27 @@
 
 #include <memory>
 
-#include "gauss2d/image.h"
-#include "pyimage.h"
+#include "lsst/gauss2d/image.h"
+#include "lsst/gauss2d/python/pyimage.h"
 
 #include <pybind11/embed.h>
 
-namespace g2 = gauss2d;
+namespace g2 = lsst::gauss2d;
 
-typedef g2::python::PyImage<double> Image; 
-typedef g2::ImageArray<double, Image> ImageArray; 
+typedef g2::python::PyImage<double> Image;
+typedef g2::ImageArray<double, Image> ImageArray;
 typedef g2::python::PyImage<bool> Mask;
 
 // Import numpy, otherwise all Image calls will segfault
 py::scoped_interpreter guard{};
 py::module_ numpy = py::module_::import("numpy");
 
-TEST_CASE("PyImage")
-{
+TEST_CASE("PyImage") {
     size_t n_rows = 3, n_cols = 2;
     std::shared_ptr<Image> image = std::make_shared<Image>(n_rows, n_cols);
     Image zeros{n_rows, n_cols};
-    for(size_t row = 0; row < n_rows; ++row) {
-        for(size_t col = 0; col < n_cols; ++col) {
+    for (size_t row = 0; row < n_rows; ++row) {
+        for (size_t col = 0; col < n_cols; ++col) {
             image->set_value(row, col, 0);
             zeros.set_value(row, col, 0);
         }
@@ -35,7 +34,7 @@ TEST_CASE("PyImage")
     CHECK(&(image->get_coordsys()) == &(g2::COORDS_DEFAULT));
     CHECK(image->get_n_cols() == n_cols);
     CHECK(image->get_n_rows() == n_rows);
-    CHECK(image->size() == n_cols*n_rows);
+    CHECK(image->size() == n_cols * n_rows);
 
     image->set_value_unchecked(0, 0, 0);
     CHECK(image->get_value_unchecked(0, 0) == 0);
@@ -44,28 +43,26 @@ TEST_CASE("PyImage")
     CHECK(*image != zeros);
     CHECK_THROWS_AS(image->get_value(0, n_cols), std::out_of_range);
     CHECK_THROWS_AS(image->set_value(n_rows, 0, 1), std::out_of_range);
-    double & value = image->_get_value_unchecked(1, 1);
+    double& value = image->_get_value_unchecked(1, 1);
     value = -1;
     CHECK(image->get_value_unchecked(1, 1) == -1);
 }
 
-TEST_CASE("VectorImageArray")
-{
+TEST_CASE("VectorImageArray") {
     size_t n_rows = 3, n_cols = 2;
     std::shared_ptr<Image> image = std::make_shared<Image>(n_rows, n_cols);
     ImageArray::Data data = {image, image};
     ImageArray arr = ImageArray(&data);
 
-    for(ImageArray::const_iterator it = arr.cbegin(); it != arr.cend(); ++it) {
+    for (ImageArray::const_iterator it = arr.cbegin(); it != arr.cend(); ++it) {
         CHECK(*it == image);
     }
-    for(const auto & img : arr) {
+    for (const auto& img : arr) {
         CHECK(img == image);
     }
 }
 
-TEST_CASE("VectorMask")
-{
+TEST_CASE("VectorMask") {
     auto image = Image(2, 2);
     auto mask = Mask(2, 2);
     CHECK(mask.get_value_unchecked(0, 0) == false);
