+++ python/lsst/__init__.py
@@ -1,5 +1,4 @@
 from pkgutil import extend_path
+
 __path__ = extend_path(__path__, __name__)
 
-from ._gauss2d import *
-from ._gauss2d import __version__
