+++ python/lsst/gauss2d/__init__.py
@@ -1,0 +1,6 @@
+from pkgutil import extend_path
+
+__path__ = extend_path(__path__, __name__)
+
+from ._gauss2d import *
+from ._gauss2d import __version__
