+++ python/lsst/gauss2d/centroid.cc
@@ -29,10 +29,12 @@
 
 #include "pybind11.h"
 
-#include "gauss2d/centroid.h"
+#include "lsst/gauss2d/centroid.h"
 
 namespace py = pybind11;
 using namespace pybind11::literals;
+
+namespace gauss2d = lsst::gauss2d;
 
 void bind_centroid(py::module &m) {
     auto _c = py::class_<gauss2d::CentroidData, std::shared_ptr<gauss2d::CentroidData>>(m, "CentroidData");
