+++ python/lsst/gauss2d/utils.py
