+++ python/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "gauss2d"
+name = "lsst-gauss2d"
 authors = [
     {name="Dan Taranu", email="dan.s.taranu@gmail.com"},
 ]
@@ -23,4 +23,10 @@ dynamic = ["version"]
 py-modules = []
 
 [tool.setuptools.dynamic]
-version = {attr = "gauss2d.__version__"}
+version = {attr = "lsst.gauss2d.__version__"}
+
+# This appears to be necessary for extend_path to find lsst in sys.modules
+[tool.setuptools.packages.find]
+where = [
+    "lsst"
+]
