+++ src/meson.build
@@ -9,4 +9,6 @@ gauss2dlib = shared_library('gauss2d',
                             gauss2d_sources,
                             include_directories : inc,
                             install : true,
+                            install_dir : get_option('libdir') / 'lsst/gauss2d',
+                            soversion : meson.project_version(),
                            )
